% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_model.R
\name{interaction_model}
\alias{interaction_model}
\title{Create a table of interaction models}
\usage{
interaction_model(data, outcome, predictor, covariates)
}
\arguments{
\item{data}{A dataframe in wide format}

\item{outcome}{A column in data with two levels to be predicted}

\item{predictor}{A column in data with two levels to predict outcome}

\item{covariates}{A vector of column names in data to model their interaction with predictor on outcome}
}
\value{
A dataframe of the table of interaction models
}
\description{
Create a table of interaction models
}
\examples{
data <- default (from ISLR package)
interaction_model(data, outcome = default, predictor = student, covariates = c("income", "balance"))
}
