% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_uppercase_fct.R
\name{my_uppercase_fct}
\alias{my_uppercase_fct}
\title{Title}
\usage{
my_uppercase_fct(x)
}
\arguments{
\item{x}{A character string}
}
\value{
The character string in uppercase
}
\description{
Title
}
\examples{
x <- "hello world"
my_uppercase_fct(x)
}
