% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_function.R
\name{return_uppercase}
\alias{return_uppercase}
\title{Title}
\usage{
return_uppercase(x)
}
\arguments{
\item{x}{A character string}
}
\value{
The character string in uppercase
}
\description{
Title
}
\examples{
x <- "hello world"
return_uppercase(x)
}
